requirements.txt
beautifulsoup4==4.7.1
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
Django==1.11.21
django-contrib-comments==1.9.1
django-cors-headers==3.0.2
djangorestframework==3.9.4
filebrowser-safe==0.5.0
future==0.17.1
grappelli-safe==0.5.2
idna==2.8
Mezzanine==4.3.1
oauthlib==3.0.1
Pillow==6.0.0
pkg-resources==0.0.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.2
tzlocal==1.5.1
urllib3==1.25.3
webencodings==0.5.1
